Copy over etc/schema/schemas.xml from Emacs
worg's org-maintenance.org says:

  schema/schemas.xml
      Any new entries in this file should be added to
      emacs/etc/schema/schemas.xml.

But that's easy to forget and instead clobber with an overwrite, as in
bug#42851.  So instead, let's keep the full version in the emacs-sync
branch.  If there are any changes on our end, there will then be merge
conflicts, and we can deal with adding the entries when merging into
emacs-sync branch rather when copying to the emacs repository.

This file is from the current emacs-27 branch (6aa9fe3e1b).