Track org-version.el
This file is needed for the emacs-sync, but if it's generated from the
emacs-sync, then the syncer needs to manually remove the hash from
org-git-version so it matches the output when the main release tag is
checked out.

Instead track it here and update the versions when the other
hard-coded versions are bumped.