ox-latex.el: Manage fonts for lualatex exports
* lisp/ox-latex: (org-latex-lualatex-font-config) Manage fonts
when exporting to lualatex.
(org-latex-lualatex-font-config): A new custom variable to manage
font configuration when exporting to to lualatex and auxiliary
functions to use it to create the fontspec configuration.
* doc/org-manual.org: Document the structure of
org-latex-lualatex-font-config and how it drives the LaTeX code it
generates.
* etc/ORG-NEWS: Announce rg-latex-lualatex-font-config

`org-latex-lualatex-font-config' is a new custom variable that stores
the configuration used by the fontspec package to configure fonts in
the output document.  It allows the user to define which fonts to use
for each document font and what features to activate for them.
Additionally, it holds a map of fallback fonts for specific scripts.
Emacs will detect them in the document and add fallbacks only for
scripts uses in the org document.