org.el: Add FONTSPEC to latex export templates
* lisp/org.el: (org-format-latex-header,org-splice-latex-header)
(org-format-latex-header): add [FONTSPEC] after the
[DEFAULT-PACKAGES].  (org-splice-latex-header): add code to treat [FONTSPEC].
* lisp/ox-latex.el: (org-latex-classes,org-latex-lualatex-fontspec-to-string)
(org-latex-classes): Add [FONTSPEC] to the docstring.
(org-latex-lualatex-fontspec-to-string): return the font specification
block as a string when `org-latex-compiler' is lualatex, else return
an empty string.  Use `org-latex--get-doc-scripts' to get the list of
scripts used in the current buffer and apply it to
`org-latex-lualatex-font-config'.
* lisp/ox-koma-letter.el: Add [FONTSPEC] to the package commentary.
* etc/ORG-NEWS: Announce [FONTSPEC] in LaTeX export templates.
* doc/org-manual.org: Document [FONTSPEC] for LaTeX export and link it
to `org-latex-lualatex-font-config'.

This series of patches inserts a reasonable font configuration using
features used by the fontspec package in the LaTeX preamble.  It has
been tested with LuaLaTeX. Other LaTeX compilers may follow.  The user
can configure the mapping between a font category (e.g. main, sans,
mono, etc.) and the font to use, adding features and potential
fallback fonts for specific scripts. Fallbacks are generated
automatically, based on the scripts detected in the current buffer.