Export fontspec for xelatex too
* etc/ORG-NEWS: Advertise new functions
`org-latex-lualatex-fontspec-to-string' and changes in
`org-splice-latex-headers'.
* lisp/org.el: (org-latex-lualatex-fontspec-to-string)
needs the font specification as a new parameter.  The font
specification needs to be generated externally in order to extract the
intended compiler from the document context.
* lisp/ox-latex.el: (org-latex-lualatex-fontspec-to-string,
org-latex-make-preamble)
(org-latex-make-preamble): add the fontspec in the call to
org-splice-latex-header.  (org-latex-lualatex-fontspec-to-string):
pass the intended latex compiler as argument.
(org-latex-lualatex-fontspec-to-string): now needs the intended
latex compiler as an argument.
lisp/ob-latex.el: (org-babel-expand-body:latex) adjust to new API for
org-splice-latex-header.

The font specification mechanism for LuaLaTeX seems to be acceptable
for XeLaTeX too, so waiting for something equivalent for PDFLaTeX to
eventually crystallise, the API has been changed, so that the intended
compiler is passed when needed and the correct header can be generated.