ox-latex.el: Omit fontspec when org-latex-lualatex-config is nil
lisp/ox-latex.el: (list-or-null-p,org-latex-lualatex-font-config,
org-latex-lualatex-fontspec-to-string)
(list-or-null-p): return non-nil when argument is a list or nil.
(org-latex-lualatex-font-config): make variable safe when it is a list
or nil using `list-or-null-p'. First attempt at defining its type for
the customisation interface.
(org-latex-lualatex-fontspec-to-string): return empty string when
`org-latex-lualatex-font-config' is nil.
doc/org-manual.org: Add that an empty fontspec will be generated
when `org-latex-lualatex-font-config' is nil.
etc/ORG-NEWS: Adverise that `org-latex-lualatex-font-config' can be
nil to omit fontspec generation with lualatex or xelatex.

When using beamer with lualatex or xelatex, the templates define
default font settings. You will normally omit any extra
fontspec configuration and go with the defaults provided by the
template.