Rename org-latex-lualatex-font-config
* lisp/ox-latex.el: (org-latex-lualatex-font-config,
org-latex-fontspec-config)
(org-latex-fontspec-config): New name for the fontspec configuration
custom variable.  It reflects the intent better.
(org-latex-lualatex-font-config): make it an alias for the new name
to keep early examples working.
* doc/org-manual.org: Refine documentation for org-latex-fontspec-config
* etc/ORG-NEWS: Reword advertisement to reflect new name

Initially all tests with the fontspec package concentrated on
lualatex.  When testing the CJK fonts example from overleaf.com on
org-mode, I discovered that the fontspec configuration mechanism could
also be used there with xelatex.  Hence the simplification of the
custom variable's name to reflect that it is *not* specific to
lualatex, but rather a way of configuring the fontspec package.