Refactor the org-tag family regex to prevent duplicates
* org.el: New constants `org-tag-valid-char-set`,
`org-tag-invalid-char-re`, `org-tag-group-enclosed-re`,
`org-tag-group-optional-re` to be used instead of inline regex.
* lisp/org.el (org-tag-re):
(org-tag-group-re):
(org-tag-line-re): Use the new regexps.
* lisp/ol-bibtex.el (org-bibtex-write):
* lisp/org-element.el (org-element--headline-parse-title):
* lisp/org.el (org-set-regexps-and-options):
(org-set-tags-command): Use the new regexp constants.

This is a refactoring patch, and there are no functional changes.  The
goal of the patch is to remove duplicates in the tag regex such that
in the future it will be easier to allow new char sets for tags.