Set org-latex-fontspec-config to nil, refine documentation
* lisp/ox-latex.el: (org-latex-fontspec-config): set default value to
nil.
(org-latex-fontspec-config): The default value used in the first
commits was conceived more as an example and was not very helpful.
Moved this initial setup to the documentation to serve as an example
and set the default value to nil for backward compatibility.
* doc/org-manual.org: Add missing references to xelatex and include
note showing how to use `org-latex-fontspec-config' for CJK fonts with
xelatex.

This change makes the feature branch be backwards compatible.  It will
not generate any fontspec configuration unless the user sets it either
in the Emacs configuration or with file or directory local
variables. The documentation now includes more examples, including the
use with xelatex for CJK handling.