Rename org-latex-lualatex-fontspec-to-string
lisp/ox-latex.el: (org-latex-fontspec-to-string,org-latex-make-preamble)
(org-latex-fontspec-to-string): renamed from
`org-latex-lualatex-fontspec-to-string' because it generates the
configuration for the fontspec package when using lualatex and
xelatex.
(org-latex-make-preamble): use the new name
lisp/ob-latex.el: (org-babel-execute:latex)
(org-babel-execute:latex): use the new name for the function.
lisp/org.el: Comment the function declaration for
`org-latex-fontspec-to-string' because it is not used in this file.
Keep it commented just in case.

Following tha same rationale as the `org-latex-fontspec-config': the
function is not exclusive for lualatex, because both lualatex and
xelatex use the fontspec package.