lisp/ox-latex.el: Fix fallback when no scripts are detected
* lisp/ox-latex.el: (org-latex-fontspec-to-string): Use `directlua' to
detect whether the fallback needs to be added to the features or not.

When `directlua' is nil, the `\\directlua{}' block was not generated,
because the font has no fallbacks or because none of the scripts in
the file ahs/needs a fallback.  In this case, don't add the fallback
to the font features.