Add tests for fontspec configuration
* testing/lisp/test-ox-latex.el:
(test-ox-latex/lualatex-fontspec-nodirectlua,
test-ox-latex/lualatex-fontspec-directlua,
test-ox-latex/lualatex-fontspec-noemoj.)

(test-ox-latex/lualatex-fontspec-nodirectlua): Check that no directlua
block is generated in the fontspec configuration does not define
fallbacks.
(test-ox-latex/lualatex-fontspec-directly): Check that directlua
block is generated when we define a fallback, and characters that use
it are present in the text.
(test-ox-latex/lualatex-fontspec-noemoji): Check that no directlua
block is generated when we define a fallback, but no characters that
use it are present in the text.

Tests for the basic functionality of the fontspec configuration
functionality using lualatex as the target compiler.