Add xpinjin and xeCJK programmatically if needed.
* lisp/ox-latex.el: (org-latex-fontspec-to-string): when CJK fonts
are present in `org-latex-fontspec-config'  and xelatex, prepend the
commands to load the xpinjin and xeCJK packages to the font
definitions derived from this variable programmatically.

The idea is not avoid having to add the packages to the
LATEX_HEADER. The packages have to be loaded before we start defining
the fonts.