Define LATEX_POLYGLOSSIA_LANGS to control the polyglossia code
* doc/org-manual.org: Document the LATEX_POLYGLOSSIA_LANGS option and
the associated `org-latex-polyglossia-languages' custom variable.
* etc/ORG-NEWS: Announce new polyglossia management options
* lisp/org.el: document modified arguments for
org-latex-fontspec-to-string.
lisp/ob-latex.el: (org-latex-fontspec-to-string, org-babel-expand-body:latex)
(org-latex-fontspec-to-string): new arguments in declaration.
(org-babel-expand-body:latex): pass polyglossia languages to
`org-latex-to-string'.
lisp/ox-latex.el: (org-latex-polyglossia-languages, org-latex-fontspec-to-string,
org-latex-make-preamble)
(org-latex-polyglossia-languages): new variable to pass the poyglossia
languages to the LaTeX header.
(org-latex-fontspec-to-string): If the `polyglossia-langs' argument is
nil, load fontspec in the document header, else load polyglossia with
the intermediate babel syntax and defer its processing to
`org-latex-guess-polyglossia-language'.
(org-latex-make-preamble): pass the compiler and the polyglossia
languages to `org-latex-fontspec-to-string'.