Pass the full info bloc to (org-latex-fontspec-to-string)
lisp/ox-latex.el: (org-latex-fontspec-to-string): use the full info
structure as parameter and get compiler, specs, etc. locally.
(org-latex-make-preamble): use the new calling convention
for (org-latex-fontspec-to-string).
lisp/ob-latex.el: (org-babel-execute:latex): change call
to (org-latex-fontspec-to-string) to create an info block.
lisp/org.el: Keep track of change in function definition

If we want to create parameters to control the language specification,
extra settings for babel/ployglossia to control the generation of the
font specification header, it s easier to extract these locally
in (org-latex-fontspec-to-string) than to pass them to the function,
taking into account that every new variable would mean yet another
parameter to (org-latex-fontspec-to-string).