lisp/ox-latex.el: Implement babel support for lualatex/xelatex
* etc/ORG-NEWS: Advertuse LATEX_BABEL_LANGS option and associated
variable (org-latex-babel-languages)
* lisp/ox-latex.el: (org-latex-babel-fontspec,
org-latex--lualatex-babel-config, org-latex-guess-babel-language):
(org-latex-babel-fontspec): new variable holding the font
specifications for the babel package.
(org-latex-guess-babel-language): Fix(simplify) detecting
`\babelprovide[..]{AUTO}'
(org-latex--lualatex-babel-config): use new variables.

`org-latex--lualatex-babel-config' will include the babel
package *without* options, and use `\babelprovide' and `\babelfont..'
to activate languages, define which of them is the main language and
associate languages and scripts to the fonts used in the document.
In the course of this development, I saw that
`org-latex-guess-babel-language' was ot detecting AUTO for
babelprovide and fixed it with simpler code.