Further simplify guessing languages for babel and add tests
* lisp/ox-latex.el: (org-latex-guess-babel-languages): Use
subexpression to replace AUTO in babelprovide[...]{AUTO} and take into
account that (replace-regexp-in-string) returns the original string
when there is no replacement.

* testing/lisp/test-ox-latex.el: (lualatex-babel-langs): New test to
check that new replacement code for babelprovide.

Using the full potential of (replace-regexp-in string)
simplifies (org-latex-guess-babel-languages) and avoids a double
search, which happened when AUTO was found in the babelprovide
statements.