Use replace-regexp-in-string, it returns the original when no replacing
* lisp/ox-latex.el: (org-latex--get-babel-lang,
org-latex--babel-langlist, org-latex-guess-babel-language):
(org-latex-guess-babel-language): use the new auxiliary functions and
`replace-regexp-in-string' to process \\usepackage[]{babel} and
\\babelprovide[]{} statements.  We create new regexps for the
replacement to isolate the subexpression we want to identify and
replace.
(org-latex--get-babel-lang): transform a short language code into a
long babel language name, processing AUTO
( org-latex--babel-langlist): transform a list of short language names
into a list of long language names, processing AUTO

The code processing the babel header can be simplified if we use the
full potential of `replace-regexp-in-string'.  This function will
return the original string if there are no replacements. We match the
complete statements and isolate the language definitions in a
subexpression.  We then replace the subexpression only.