lisp/ox-latex.el: Generate polyglossia a priori, don't guess
* lisp/ox-latex.el (org-latex--polyglossia-fonts,
org-latex-guess-polyglossia, org-latex--lualatex-polyglossia-config,
org-latex-fontspec-to-string, org-latex-make-preamble):
(org-latex--polyglossia-fonts): mark as obsolote.
(org-latex-gues-polyglossia): mark as obsolote.
(org-latex--lualatex-polyglossia-config): Generate the polyglossia
import and font specification in this function instead of guessing and
replacing.
(org-latex-fontspec-to-string): Call (org-lualatex-polyglossia-config)
when there is polyglossia configuration but no babel configuration in
the document.
(org-latex-make-preamble): no need to guess anything related with
polyglossia. We are generating the full spec when required.

This is part of the rewrite of the font support.  It is easier to
settle on a specific scheme depending on things like the compiler and
define the multi-language options in the document and then to generate
the full header, than generating a stub and then replacing things in
the header a posteriori.  Currently, polyglossia and babel (for
lualatex) are generated separately.  The code defaults to fontspec if
none is selected.