Preparing for the future: org-latex-multi-lang-driver
* lisp/ox-latex.el: (org-latex-multi-lang-driver)
(org-latex-multi-lang-driver) Set this variable to `t' to activate the
new font management and multi-language support variables.
* testing/lisp/test-ox-latex.el: (test-ox-latex/lualatex-babel-langs)
(test-ox-latex/lualatex-babel-langs): adjust this test to use LANGUAGE
and LATEX_MULTI_LANG instead of LATEX_BABEL_LANGS with AUTO.

This step should make it easier to isolate "legacy" behaviour based on
usepackage and AUTO to the new implementation, where LANGUAGE holds
the list of languages present in the document and LATEX_MULTI_LANG
activates the "backend" package (babel, polyglossia) or just the new
fontspec-based generator.