A proof-of-concept for integrating new features
* lisp/ox-latex.el: (org-latex--get-babel-lang,
org-latex--lualatex-babel-config)
(org-latex--get-babel-lang): Make default-lang parameter optional.
When it is missing, "AUTO" as language produces an error.
(org-latex--lualatex-babel-config): Use new calling convention when
getting the babel language to catch people using AUTO in languages,
which should not be the case.  LANGUAGES can use short names.

The idea is to isolate the differences in the work-flow for new
vs. old features.  It clarifies when is new code while keeping backward
compatibilty.  Should help integrate the new code.