First attempt at pdflatex + babel
* list/ox-latex.el: (org-latex--lualatex-polyglossia-config,
org-latex--pdflatex-fontconfig,org-latex--pdflatex-ldf,
org-latex--pdflatex-encode)
(org-latex--lualatex-polyglossia-config): cosmetic change
(org-latex--pdflatex-fontconfig): Generate code for babel+pdflatex and
pdflatex only.
(org-latex--pdflatex-ldf): Auxiliary function to derive the babel
language name for pdflatex. Only takes :babel prop into
account. Generates error if language doesn't have this property.
(org-latex--pdflatex-encode): In embrionary state, returns LGR for
greek and T2A for russian. The rest of languages need to be added in
the future!

The objective was to generate a LaTeX header that can generate some
multi-language documents as a proof of concept.