Fix polyglossia bug reported by Christian Moe
* lisp/org.el: (org-splice-latex-header): Move FONTSPEC replacement
after EXTRA replacement.
* lisp/ox-latex.el: (org-latex--lualatex-polyglossia-config): get
polyglossia language name from (org-latex-languages-alist).

Christan Moe bissected a bug when exporting to xelatex. The font names
were not correct. When fixing this, a new error appreared when using
bidi features. This implied that importing the multi-language handling
packages and defining fonts should be the last thing to do.