Silence hyperref error for multiple languages
* lisp/ox-latex.el: (org-latex--get-lang, org-latex--format-spec)
(org-latex--get-lang): factor out the language related code. Split the
language string and keep the first language.
(org-latex--format-spec): Use (org-latex--get-lang) to manage the
languages.

If we have multiple languages in the pdflang spec for hyperref, we get
a warning. TODO: as per hyperref spec, the language should be an ISO
code.