lisp/ox-latex.el: Allow nil in org-latex-babel-fontspec as language
* lisp/ox-latex.el: (org-latex--lualatex-polyglossia-config,
org-latex--lualatex-babel-config)

(org-latex--lualatex-polyglossia-config): Use (org-latex--mk-options)
to build set..font
(org-latex--lualatex-babel-config): Use (org-latex--mk-options)
instead of "[%s]". Allow `nil' as language in
`org-latex-babel-fontspec' to build `\babelfont{fam}{font}'

With this change in  org-latex--lualatex-babel-config we can define
babel fonts common to all languages in the document by associating  nil
instead of an Org language code to a set of fonts.