lisp/ox-latex.el: provide a default font set for jp+babel+xelatex
lisp/ox-latex.el: (jp-default-fontspec): a new alist that maps the CJK
fonts to their HaranoAji counterpart for Japanese documents.
(org-latex--utf8latex-babel-config): when the main language is "jp",
check (org-latex-fontspec-config). It if is nil or does not contain
mappings for CJK fonts, use (jp-default-fontspec) to define the CJK
fonts.

The rationale: zh+babel+xelatex doesn't need an explicit font
specification. We provide one for jp+babel+xelatex that follows the
convention for Japanese documents, so no explicit font specification
is needed here either.