ob-core.el: Fix incorrect interposition of Noweb prefixes
* lisp/ob-core.el (org-babel-expand-noweb-references): Fix incorrect
interposition of Noweb prefixes.

Currently, the following example

,#+name: a string
,#+begin_src emacs-lisp
"hello"
,#+end_src

,#+begin_src emacs-lisp :tangle tangle-bug.el :noweb yes :comments noweb
prefix<<a string>>
,#+end_src

tangles to

,#+begin_src emacs-lisp
;; [[file:org-babel-tangle.org::*Incorrect interposition of Noweb prefix when =:comments noweb=][Incorrect interposition of Noweb prefix when =:comments noweb=:2]]
prefix;; [[file:org-babel-tangle.org::a string][a string]]
prefix"hello"
prefix;; a string ends here
prefix
;; Incorrect interposition of Noweb prefix when =:comments noweb=:2 ends here
,#+end_src

where we get one more occurrence of "prefix" than expected at the end of
the expansion.  This is because org-babel-expand-noweb-references adds
an empty line after the Noweb comments blindly, *before* interposition
happens.

In this patch, the empty line is only added when needed (i.e., when
Noweb comments are desired, and the block body does not end with a Noweb
reference), and is done after the interposition.

* testing/lisp/test-ob.el: Add test.

Reported-by: "Mingtong Lin" <mt.oss@fastmail.com>
Link: https://list.orgmode.org/f43360bb-dc8f-41bb-b40e-dfdd38ebb87b@app.fastmail.com/