lisp/org-mobile.el: Fix numeric priorities with org-mobile
* lisp/org.el (org-priority-value-regexp, org-priority-regexp,
org-set-regexps-and-options): Move `org-priority-value-regexp' and
`org-priority-regexp' alongside other early regexps to prevent a "reference
to free variable" warning.  Update `org-set-regexps-and-options' to use the
existing `org-priority-value-regexp' for the priority value pattern in the
`org-complex-heading-regexp' and `org-complex-heading-regexp-format'
regexps so that all valid priority values are matched.

* lisp/org-mobile.el (org-mobile-edit): Fix handling of numeric priorities
by parsing with `org-priority-to-value' and using the full priority
substring from the cookie for the current value.

* testing/lisp/test-org-mobile.el: Add a new file for unit tests related to
the `org-mobile' functionality.  Add a unit test to exercise handling of
numeric priorities with the `org-mobile-edit' function.